Camera rotation code moved to Camera script
The player object was being rotated on the X-axis while the camera
was being rotated on the Y-axis to simulate the look-around. I've
combined the moves into Spatial's rotate_object_local calls so
that the Camera can do all the rotation work while the player
object handles only movement.

Reticle's mouse filter set to MOUSE_FILTER_IGNORE

Player's _unhandled_input was not triggering because
MOUSE_FILTER_PASS, the original setting, handles the
input even when it isn't grabbed by a parent.