GFX update, & decal moved into weapon script
Made a rough pass at some scorchmarks for the decal, and since the
FPS is kind of sci-fi-y, made a sci-fi-y crosshair. Old files
removed.

Moved the decal generation script into a Node in the Player scene
and out of Demo.gd, since it's a function of the weapon shooting
rather than the game.